+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - console: add create table functionality for MS SQL Server 
 - console: update connect database form with SSL certificates
 - console: add drop table functionality to MS SQL Server tables
+- console: allow renaming data sources
 
 ## v2.0.0-alpha.11